Move remarkable plugins to website/lib
* Bypass escaped delimiter
