+++ website/siteConfig.js
@@ -60,7 +60,10 @@ const siteConfig = {
         theme: 'default',
     },
 
-    markdownPlugins: [require('./remarkableKatex'), require('./empty_thead')],
+    markdownPlugins: [
+        require('./lib/remarkableKatex'),
+        require('./lib/empty_thead'),
+    ],
 
     scripts: ['https://buttons.github.io/buttons.js'],
     stylesheets: ['https://cdn.jsdelivr.net/npm/katex@0.10.0-beta/dist/katex.min.css'],
